dynamic log entry sizes | optimize conditions for reqVote request handler

+++ lib/rpc/messages.h
@@ -53,7 +53,7 @@ public:
 
     uint32_t term;
     uint32_t candidateID;
-    uint8_t lastLogIndex;
+    uint16_t lastLogIndex;
     uint32_t lastLogTerm;
 
     /**
